+++ java/src/hype/core/util/HWarnings.java
@@ -33,6 +33,9 @@ public class HWarnings {
 		CHILDCEPTION =
 		"Can't add this parent as its own child.",
 		
+		INVALID_CHILD =
+		"The child you're trying to add is cannot be added to this drawable.",
+		
 		VERTEXPX_ERR =
 		"Set a non-zero size first for this path before setting the\n\t" +
 		"vertex by pixels, or use the vertexPerc() methods instead.";